pyinfra/pyproject.toml
[tool.poetry]
name = "pyinfra-ineedempathy"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-digitalocean = "^1.16.0"
pyinfra = "^1.4.12"
invoke = "^1.6.0"

[tool.poetry.dev-dependencies]
pdbpp = "^0.10.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
